requirements.txt
setuptools
wheel
multidict
numpy
python-dateutil
ax-platform
art
tzlocal
Rich
rich-argparse
sixel
scikit-learn
submitit
matplotlib
tqdm
seaborn
pytz
psutil
coverage
pyflakes
cowsay
beautifulsoup4
emoji
hiplot
html5lib
lxml
toml
rich_argparse
beartype
pyyaml
pyspellchecker
pyfiglet
setuptools
pyright
